Stage 3 of 3 for JetBrains Academy - Kotlin Android - Tip Calculator project. Description
You're almost there! In the final stage of the project, you need to take the user input data and perform the calculations. To determine the tip amount, take the overall bill value, multiply the number by the tip percentage, and divide it by 100. Output the tip amount with two digits in the fraction part.
Objectives

Using the event listeners from the previous stage, update the tip calculation result in the TextView each time the user changes the bill value or tip percentage.

MainActivity.kt changes:
• adjusted the slider listener to return the value as Double instead of Int.

+++ Tip Calculator/task/src/main/java/org/hyperskill/calculator/tip/MainActivity.kt
@@ -24,7 +24,7 @@ class MainActivity : AppCompatActivity() {
         }
 
         findViewById<Slider>(R.id.slider).addOnChangeListener { _, value, _ ->
-            calculator.percentChange(value.toInt())
+            calculator.percentChange(value.toDouble())
         }
     }
 }
